Separated components into different files
